PWGJE: separating different devices from the same workflow
